backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
yt-dlp==2023.11.16
openai-whisper==20231117
pydub==0.25.1
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
cors==1.0.1
websockets==12.0